requirements.txt
# local package
-e .

# external requirements
boto3==1.34.79
click==8.1.7
geopandas==0.14.3
fastapi==0.110.1
ipykernel==6.29.4
matplotlib==3.8.4
netCDF4==1.6.5
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0
rasterio==1.3.9
tqdm==4.66.2
uvicorn==0.29.0
xarray==2024.3.0